changed indentation for the RETURN STATUS_ERROR code
I found the return STATUS_ERROR somewhat obstructs the view
on what is really going on. Thus I tried to make the actual
code a little bit more prominent. The price to pay is to
somewhat deviate from the overall indentation style.

However this is a matter of personal taste. If you like it
merge it, otherwise reject it.