bugfix cache handling vs conversion
Cache handling had several issues:

1) The old cache had dummy reads in the readData function which filled
   the cached attributes as a side effect. The new cache did NOT
   create this side effect.

2) The old as well as the new cache did not take into account that
   convertRaw was exposed independently of the readData function.

3) Both approaches did not take into account that fill_cache will
   always fails if burst mode is already active before the first
   call to readData. The only way to mitigate this is to ensure the
   cache is already filled before anyone might start burst mode.

The fix ensures that the cache is reloaded after all functions
might invalidate it. Unfortunately this is not 100% waterproof.
The issue is that reset does not clear the burst flag. Thus the cache
will not fill properly after reset.

Mitigation: exit() should always be called before calling reset()