improved cache handling
The new approach will not try to get the cache correct as early as possible.
Instead it will follow the strategy below:

1) Invalidate it whenever a command would have side effects that might
get it out of sync.

2) Force to fill it prior to commands that start measurement.

This is more consistent and avoids refilling the cache e.g. after reset.
The point is that filling the cache after a reset can be impossible if
the sensor is still in burst mode.

The overall construction is still brittle if someone play around with
burst mode.

The proper way to deal with burst mode is to ALWAYS exist it prior to
a reset. Then everything should work as intended. However I did not
want to encode this into the library as it might break other peoples
code in unexpected ways.